requirements.txt
# Core dependencies
numpy>=1.26.4,<2.0.0
pandas>=2.2.1
PyQt5>=5.15.10
python-dotenv>=1.0.1

# AI and ML dependencies
transformers>=4.38.2
mistralai>=0.3.0  # Add Mistral AI dependency

# Vector database
chromadb>=0.4.22

# LLM indexing
llama-index>=0.10.13

# SwarmZero
swarmzero==0.0.1

# Async support
aiohttp>=3.9.3
asyncio>=3.4.3
nest-asyncio>=1.5.9

# Utilities
tqdm>=4.66.2
requests>=2.31.0
python-dateutil>=2.8.2
qasync==0.27.1

# livepeer
livepeer-ai==0.5.2
livepeer==0.5.2
